requirements.txt
Django>=5.2.4
Pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.5.0
openai
google-generativeai
huggingface_hub
python-dotenv
numpy
requests
